Struct bracket_geometry::prelude::Rect

pub struct Rect {
    pub x1: i32,
    pub x2: i32,
    pub y1: i32,
    pub y2: i32,
}

Defines a two-dimensional rectangle.

Fields
x1: i32

The X position of the first point (typically the left)

x2: i32

The X position of the second point (typically the right)

y1: i32

The Y position of the first point (typically the top)

y2: i32

The Y position of the second point (typically the bottom)

Implementations

impl Rect

pub fn with_size<T>(x: T, y: T, w: T, h: T) -> Rect
where T: TryInto<i32>,

Create a new rectangle, specifying X/Y Width/Height

pub fn with_exact<T>(x1: T, y1: T, x2: T, y2: T) -> Rect
where T: TryInto<i32>,

Create a new rectangle, specifying exact dimensions

pub fn zero() -> Rect

Creates a zero rectangle

pub fn intersect(&self, other: &Rect) -> bool

Returns true if this overlaps with other

pub fn center(&self) -> Point

Returns the center of the rectangle

pub fn point_in_rect(&self, point: Point) -> bool

Returns true if a point is inside the rectangle

pub fn for_each<F>(&self, f: F)
where F: FnMut(Point),

Calls a function for each x/y point in the rectangle

pub fn point_set(&self) -> HashSet<Point>

Gets a set of all tiles in the rectangle

pub fn width(&self) -> i32

Returns the rectangle’s width

pub fn height(&self) -> i32

Returns the rectangle’s height